Refactoring directories & cleaning & moving files
