+++ BusinessAssistant-login/src/main/java/com/businessassistantbcn/login/config/PropertiesConfig.java
@@ -1,4 +1,4 @@
-package com.businessassistantbcn.login;
+package com.businessassistantbcn.login.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @EnableConfigurationProperties
 @ConfigurationProperties("security.datasource")
-public class YAMLConfig {
+public class PropertiesConfig {
 	
 	private String signUpUrl;
 	private String secret;
