+++ BusinessAssistant-login/src/main/java/com/businessassistantbcn/login/controller/LoginController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/v1")
+@RequestMapping("/v1/api")
 public class LoginController {
 
     @Autowired
@@ -17,7 +17,7 @@ public class LoginController {
         return "Hello from BusinessAssistant Barcelona!!!";
     }
     
-    @PostMapping("/api/login")
+    @PostMapping("/login")
     public String test2() {
         return "Login functional";
     }
