+++ BusinessAssistant-opendata/build.gradle
@@ -15,8 +15,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web:2.5.6'
     implementation 'org.springframework.boot:spring-boot-starter-webflux:2.6.1'
 
-    implementation 'org.springframework:spring-tx:version:5.3.12'
-    implementation 'org.springframework.data', name: 'spring-data-jpa', version: '2.5.6'
     implementation 'io.springfox:springfox-boot-starter:3.0.0'
     implementation 'io.springfox:springfox-swagger-ui:3.0.0'
     implementation 'org.json:json:20210307'
