+++ BusinessAssistant-opendata/src/main/java/com/businessassistantbcn/opendata/helper/HttpClientHelper.java
@@ -3,5 +3,5 @@ package com.businessassistantbcn.opendata.helper;
 import org.springframework.stereotype.Component;
 
 @Component
-public class HttpClient {
+public class HttpClientHelper {
 }
