--- postman/BABCN-4 JWT Test.postman_collection.json
@@ -1,155 +1,0 @@
-{
-	"info": {
-		"_postman_id": "976f916a-f117-49bc-aa77-86db0ba0387d",
-		"name": "BABCN-4 JWT Test",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "http://localhost:8761/v1/test (sin token)",
-			"event": [
-				{
-					"listen": "test",
-					"script": {
-						"exec": [
-							"pm.test(\"Token absent - Status code is 401 (UNAUTHORIZED)\", function () {\r",
-							"    pm.response.to.have.status(401);\r",
-							"});"
-						],
-						"type": "text/javascript"
-					}
-				}
-			],
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8761/v1/test",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8761",
-					"path": [
-						"v1",
-						"test"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "http://localhost:8761/v1/test (con token válido)",
-			"event": [
-				{
-					"listen": "test",
-					"script": {
-						"exec": [
-							"pm.test(\"Status code is 200\", function () {\r",
-							"    pm.response.to.have.status(200);\r",
-							"});"
-						],
-						"type": "text/javascript"
-					}
-				}
-			],
-			"request": {
-				"method": "GET",
-				"header": [
-					{
-						"key": "Authorization",
-						"value": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyLCJleHAiOjk5OTk5OTk5OTl9.x38rH6CwepOXV3yK_YcGY8eP85gL1cvNKTJ4EBv5Vvw",
-						"type": "text"
-					}
-				],
-				"url": {
-					"raw": "http://localhost:8761/v1/test",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8761",
-					"path": [
-						"v1",
-						"test"
-					],
-					"query": [
-						{
-							"key": "Authorization",
-							"value": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
-							"disabled": true
-						}
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "http://localhost:8761/v1/test (con token caducado)",
-			"event": [
-				{
-					"listen": "test",
-					"script": {
-						"exec": [
-							"pm.test(\"Expired token - Status code is 401 (UNAUTHORIZED)\", function () {\r",
-							"    pm.response.to.have.status(401);\r",
-							"});"
-						],
-						"type": "text/javascript"
-					}
-				}
-			],
-			"request": {
-				"method": "GET",
-				"header": [
-					{
-						"key": "Authorization",
-						"value": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkNoaXF1aXRvIGRlIGxhIENhbHphZGEiLCJpYXQiOjAsImV4cCI6NjkzNzAyMDAwfQ.Q99Jbns7GmNYTVylVDOYRBff7RyUUVhgJoiCfuwyaD0",
-						"type": "text"
-					}
-				],
-				"url": {
-					"raw": "http://localhost:8761/v1/test",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8761",
-					"path": [
-						"v1",
-						"test"
-					],
-					"query": [
-						{
-							"key": "Authorization",
-							"value": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
-							"disabled": true
-						}
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "http://localhost:8761/v1/api/login",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8761/v1/api/login",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8761",
-					"path": [
-						"v1",
-						"api",
-						"login"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
