--- postman/BusinessAssistant-opendata.postman_collection.json
@@ -1,141 +1,0 @@
-{
-	"info": {
-		"_postman_id": "f0df0311-254f-4353-94e9-6921d04ad2e4",
-		"name": "BusinessAssistant-opendata",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "test",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/test",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"test"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "large-establishments",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/large-establishments",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"large-establishments"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "commercial-galeries",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/commercial-galeries",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"commercial-galeries"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "big-malls",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/big-malls",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"big-malls"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "municipal-markets",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/municipal-markets",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"municipal-markets"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "markets-fairs",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8762/v1/api/opendata/markets-fairs",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8762",
-					"path": [
-						"v1",
-						"api",
-						"opendata",
-						"markets-fairs"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
